+++ erc20transfer/schema.graphql
@@ -74,6 +74,7 @@ type Transfer @entity(immutable: true) {
   blockNumber: BigInt!
   blockTimestamp: BigInt!
   transactionHash: Bytes!
+  tokenAddress: Bytes # address
 }
 
 type Pause @entity(immutable: true) {
@@ -87,38 +88,6 @@ type Pause @entity(immutable: true) {
 type Unpause @entity(immutable: true) {
   id: Bytes!
 
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
-
-type daiApproval @entity(immutable: true) {
-  id: Bytes!
-  src: Bytes! # address
-  guy: Bytes! # address
-  wad: BigInt! # uint256
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
-
-type LogNote @entity(immutable: true) {
-  id: Bytes!
-  sig: Bytes! # bytes4
-  usr: Bytes! # address
-  arg1: Bytes! # bytes32
-  arg2: Bytes! # bytes32
-  data: Bytes! # bytes
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
-
-type daiTransfer @entity(immutable: true) {
-  id: Bytes!
-  src: Bytes! # address
-  dst: Bytes! # address
-  wad: BigInt! # uint256
   blockNumber: BigInt!
   blockTimestamp: BigInt!
   transactionHash: Bytes!
