+++ erc20transfer/src/erc20.ts
@@ -1,0 +1,19 @@
+import { Transfer as TransferEvent } from "../generated/usdt/usdt";
+import { Transfer } from "../generated/schema";
+
+export function handleTransfer(event: TransferEvent): void {
+  let entity = new Transfer(
+    event.transaction.hash.concatI32(event.logIndex.toI32())
+  );
+  entity.from = event.params.from;
+  entity.to = event.params.to;
+  entity.value = event.params.value;
+
+  entity.blockNumber = event.block.number;
+  entity.blockTimestamp = event.block.timestamp;
+  entity.transactionHash = event.transaction.hash;
+
+  entity.tokenAddress = event.transaction.to;
+
+  entity.save();
+}
