--- erc20transfer/tests/usdt.test.ts
@@ -1,46 +1,0 @@
-import {
-  assert,
-  describe,
-  test,
-  clearStore,
-  beforeAll,
-  afterAll
-} from "matchstick-as/assembly/index"
-import { BigInt, Address } from "@graphprotocol/graph-ts"
-import { Issue } from "../generated/schema"
-import { Issue as IssueEvent } from "../generated/usdt/usdt"
-import { handleIssue } from "../src/usdt"
-import { createIssueEvent } from "./usdt-utils"
-
-// Tests structure (matchstick-as >=0.5.0)
-// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0
-
-describe("Describe entity assertions", () => {
-  beforeAll(() => {
-    let amount = BigInt.fromI32(234)
-    let newIssueEvent = createIssueEvent(amount)
-    handleIssue(newIssueEvent)
-  })
-
-  afterAll(() => {
-    clearStore()
-  })
-
-  // For more test scenarios, see:
-  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test
-
-  test("Issue created and stored", () => {
-    assert.entityCount("Issue", 1)
-
-    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
-    assert.fieldEquals(
-      "Issue",
-      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
-      "amount",
-      "234"
-    )
-
-    // More assert options:
-    // https://thegraph.com/docs/en/developer/matchstick/#asserts
-  })
-})
